Update and rename ProductDeletion.java to ProductUtil.java
